Moved swapping projections into map class.
